moved derived metrics tests out of CFFI module
